refactor(discover_schedule): 重构定时发现任务模块
重构定时发现任务模块，包括以下改进：
1. 将discover_task重命名为discover_schedule以更准确描述功能
2. 优化代码结构，调整import顺序
3. 添加完整的CRUD操作实现
4. 增加定时任务执行逻辑
5. 修复测试用例中的mock引用
6. 移除冗余的空行和注释
