+++ adp/vega/vega-backend/server/drivenadapters/discover_schedule/discover_schedule_access.go
@@ -3,7 +3,7 @@
 // Licensed under the Apache License, Version 2.0.
 // See the LICENSE file in the project root for details.
 
-package discover_task
+package discover_schedule
 
 import (
 	"context"
