+++ adp/vega/vega-backend/server/drivenadapters/kafka/kafka_access.go
@@ -3,7 +3,7 @@
 // Licensed under the Apache License, Version 2.0.
 // See the LICENSE file in the project root for details.
 
-package drivenadapters
+package kafka
 
 import (
 	"context"
