+++ adp/vega/vega-backend/server/driveradapters/resource_handler.go
@@ -587,4 +587,3 @@ func (r *restHandler) ListResources(c *gin.Context) {
 	}
 
 }
-
