+++ adp/vega/vega-backend/server/logics/connectors/local/index/opensearch/opensearch_condition.go
@@ -9,6 +9,7 @@ import (
 	"fmt"
 	"strings"
 	"time"
+
 	"vega-backend/interfaces"
 	"vega-backend/logics/filter_condition"
 )
