+++ adp/vega/vega-backend/server/logics/query/stream_query_manager.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/kweaver-ai/kweaver-go-lib/logger"
+
 	"vega-backend/interfaces"
 )
