+++ adp/vega/vega-backend/server/logics/resource/resource_service.go
@@ -764,4 +764,3 @@ func (rs *resourceService) CheckExistByCategories(ctx context.Context, catalogID
 
 	return rs.ra.CheckExistByCategories(ctx, catalogID, categories)
 }
-
